Add missing EqualsAndHashCode annotation to User model
The existing poor implementation of User.hashCode caused
AuthController.register tests to fail, so replacing the hashCode method
with a proper one and its missing equals method made the tests pass.

Apparently, the missing EqualsAndHashCode annotation is a recurring
theme in our model classes, causing all sorts of unexpected bugs in
other parts of the project.